src/requirements.txt
aiofiles==0.5.0
aiohttp==3.7.4.post0
airtable-python-wrapper==0.15.1
aniso8601==7.0.0
astroid==2.5.1
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
discord.py==1.6.0
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
orjson==3.5.1
promise==2.3
pydantic==1.8.1
pylint==2.7.2
python-dotenv==0.15.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
Rx==1.6.1
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.3
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
wrapt==1.12.1
yarl==1.6.3
